refactor(multitable): extract record query service
