+++ docs/development/multitable-m3-query-service-verification-20260423.md
@@ -1,0 +1,55 @@
+# Multitable M3 query-service extraction verification
+
+Date: 2026-04-23
+Branch: `codex/multitable-m3-query-service-20260423`
+Base: `origin/main@6a677f9c3`
+
+## Commands
+
+Run from worktree root:
+
+```bash
+pnpm install --prefer-offline
+pnpm --filter @metasheet/core-backend exec vitest run \
+  tests/unit/multitable-query-service.test.ts \
+  tests/unit/multitable-records.test.ts \
+  tests/unit/multitable-cursor-pagination.test.ts \
+  tests/unit/multitable-plugin-scope.test.ts \
+  --reporter=dot
+pnpm --filter @metasheet/core-backend exec tsc --noEmit
+```
+
+## Results
+
+Focused test run:
+
+```text
+Test Files  4 passed (4)
+Tests       36 passed (36)
+```
+
+Breakdown:
+
+- `multitable-query-service.test.ts`: `4/4` passed.
+- `multitable-records.test.ts`: `13/13` passed.
+- `multitable-cursor-pagination.test.ts`: `12/12` passed.
+- `multitable-plugin-scope.test.ts`: `7/7` passed.
+
+TypeScript:
+
+- `pnpm --filter @metasheet/core-backend exec tsc --noEmit`: exit `0`.
+
+## Coverage Notes
+
+- New direct query-service coverage verifies cursor/cache helpers, list
+  delegation, filter/search/order SQL generation, and cursor pagination metadata.
+- Existing `records.ts` tests remain green, proving compatibility for imports
+  that still consume `multitable/records.ts`.
+- Existing plugin-scope tests remain green, proving plugin API callers are not
+  forced to move to the new module.
+
+## Local Environment Note
+
+`pnpm install --prefer-offline` was required because this newly created worktree
+did not have workspace dependency links yet. The resulting `node_modules` files
+are install artifacts and were not staged.
