+++ packages/core-backend/src/multitable/record-errors.ts
@@ -1,0 +1,7 @@
+export class MultitableRecordValidationError extends Error {
+  code = 'VALIDATION_ERROR'
+}
+
+export class MultitableRecordNotFoundError extends Error {
+  code = 'NOT_FOUND'
+}
